// src/store/filtersStore/interfaces.ts
export enum Filters {
  BI = 'BI-отчеты',
  GOS = 'Гос. заказ',
  ENERGY = 'Энергетика',
  EDUCATION = "Образование",
  FIN_TECH = 'Финтех',
  LIFESTYLE = 'Образ жизни',
  MEDICINE = 'Медицина',
  E_COMMERCE = 'E-commerce',
  SAFETY = 'Безопасность',
  ENTERTAINMENTS = 'Развлечения',
  CARTOGRAPHY = 'Картография',
  
  ANDROID = 'Android', 
  CHAT_SYSTEM = 'Chat-system',
  TELEGRAM = 'Telegram',
  BACKEND = 'Backend',
  WINDOWS = 'Windows', 
  CONTROL_PANEL = 'Панель управления',
  WEB = 'Web',
  IOS = 'iOS',

  RELEASE = 'Релиз',
  SUPPORT = 'Сопровождение',
  DESIGN = 'Дизайн',
  TESTING = 'Тестирование',
  DEVELOPMENT = 'Разработка',
  PRE_PROJECT_ACTIVITIES = 'Предпроектная деятельность',
  ANALYTICS = 'Аналитика',
  ACCEPTANCE = 'Приемка',

  NODE = 'Node.JS',
  JAVA = 'Java',
  XAMARIN = 'Xamarin',
  WORDPRESS = 'Wordpress',
  DRUPAL_ANGULAR = 'Drupal 8, Angular',
  NOT_AZURE = 'ASP.NET Core, Azure',
  MS_POWER_BI = 'MS Power BI',
  BITRIX = '1C-Битрикс',
  OBJECTIVE_C = 'Objective-C',
  C_SHARP_UWP = 'C#, UWP',
  PYTHON = 'Python',
  PHP = 'PHP',
  KOTLIN = 'Kotlin',
  MODX = 'MODX',
  SWIFT = 'Swift',
  PHONEGAP = 'PhoneGap',
  REACT = 'React',
  JAVA_KOTLIN = 'Java, Kotlin',
  CORDOVA = 'Cordova',
  C_PLUS = 'C++'
}